+++ src/apps.js
@@ -2,12 +2,13 @@ import fs from "fs";
 import path from "path";
 import fetch from "node-fetch";
 import AdmZip from "adm-zip";
+import { exec } from "child_process";
 
 async function downloadZip(url, destination) {
   const response = await fetch(url);
 
   if (!response.ok) {
-    throw new Error(`Failed to fetch the file: ${response.statusText}`);
+    throw new Error(response.statusText);
   }
 
   const buffer = await response.arrayBuffer();
@@ -16,12 +17,44 @@ async function downloadZip(url, destination) {
   return destination;
 }
 
+async function downloadFile(url, destination) {
+  const response = await fetch(url);
+
+  if (!response.ok) {
+    throw new Error(response.statusText);
+  }
+
+  const buffer = await response.arrayBuffer();
+  fs.writeFileSync(destination, Buffer.from(buffer));
+  return destination;
+}
+
 export const Apps = {
   async install(id, repo) {
     const homeDir = process.env.HOME || process.env.USERPROFILE;
     const mrserverDir = path.join(homeDir, ".mrserver");
     const zipPath = path.join(mrserverDir, "app.zip");
-
+    try {
+      await downloadFile(`${repo}/${id}/preinstall.json`, path.join(mrserverDir, "preinstall.json"));
+      console.log("💻 [MrServer CLI] Executing preinstall script...");
+      const preinstall = JSON.parse(fs.readFileSync(path.join(mrserverDir, "preinstall.json"), "utf8"));
+      if (preinstall) {
+        await exec(`${preinstall.shell} ${preinstall.shellFlags} "${preinstall.command}"`, (error, stdout, stderr) => {
+          if (error) {
+            console.error(`❌ [MrServer CLI] Error executing preinstall script: ${error.message}`);
+            throw error;
+          }
+          if (stderr) {
+            console.error(stderr);
+            throw new Error(stderr);
+          }
+          console.log(stdout);
+        });
+        fs.unlinkSync(path.join(mrserverDir, "preinstall.json"));
+      }
+    } catch {
+      console.log("ℹ️  [MrServer CLI] No preinstall script found.. continuing with installation.");
+    }
     await downloadZip(`${repo}/${id}/app.zip`, zipPath);
     console.log("✅ [MrServer CLI] Installing app...");